Use shared view model with simple payment dialog
